refactor: move src/Preview.tsx to src/components/preview.tsx
